Move vulpix to `dotty.tools.vulpix`
